Rename open contract to bought contract
