+++ src/_reducers/BoughtContractsReducer.js
