--- src/_reducers/ContractsReducer.js
@@ -1,22 +1,0 @@
-import { fromJS } from 'immutable';
-
-import {
-    SERVER_DATA_PORTFOLIO,
-    REMOVE_PERSONAL_DATA,
-} from '../_constants/ActionTypes';
-
-const initialState = fromJS([]);
-
-export default (state = initialState, action) => {
-    switch (action.type) {
-        case SERVER_DATA_PORTFOLIO: {
-            const contracts = action.serverResponse.portfolio.contracts;
-            return state.merge(fromJS(contracts));
-        }
-        case REMOVE_PERSONAL_DATA: {
-            return initialState;
-        }
-        default:
-            return state;
-    }
-};
