+++ src/portfolio/PortfolioSelectors.js
@@ -1,16 +1,16 @@
-import { openContractsSelector, portfolioSelector, ticksSelector } from '../_store/directSelectors';
+import { boughtContractsSelector, portfolioSelector, ticksSelector } from '../_store/directSelectors';
 import { createSelector, createStructuredSelector } from 'reselect';
 
 const activeOpenContractSelector = createSelector(
-    openContractsSelector,
-    openContracts =>
+    boughtContractsSelector,
+    boughtContracts =>
         /**
          * check 'contract_id' is a synchronization, as contracts is combination of 2 data point,
          * transaction_id is from transactions stream, the rest from open contract stream
          * this check ensure open contract response is received before rendering,
          * but does not ensure `transaction_id` is sync
          */
-        openContracts
+        boughtContracts
             .filter(c => !c.get('sell_price') && c.get('contract_id'))
 );
